requirements.txt
requests
together
deep_translator
python-dotenv
cachetools
ratelimit
